refactor: 회원가입 API auth -> user 경로 변경 (#3) - sign-up API를 AuthController → UserController로 이동
